# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /patients/{id}/problems:
    get:
      operationId: getPatientProblems
      tags: [coding]
      summary: Local problem-list cache for a patient
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Problem list
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/PatientProblem" }
  /coding/sessions/{id}/problem-suggestions:
    get:
      operationId: getProblemSuggestionsForSession
      tags: [coding]
      summary: Reconciler output for a Coder session
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Problem-list suggestions
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/ProblemListSuggestion" }
  /coding/sessions/{id}/reconcile-problems:
    post:
      operationId: reconcileProblemsForSession
      tags: [coding]
      summary: Re-run problem-list reconciliation against this session
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Reconciliation produced (or replaced) suggestions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReconcileResponse"
        "404":
          description: Session or patient not found
        "409":
          description: Note has no assessment section to reconcile against
  /problem-list-suggestions/{id}/accept:
    post:
      operationId: acceptProblemListSuggestion
      tags: [coding]
      summary: Accept a problem-list reconciliation action (applies the local mutation)
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AcceptProblemListSuggestionRequest" }
      responses:
        "200":
          description: Accepted; local mutation applied
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProblemListSuggestion" }
        "404":
          description: Suggestion not found
        "409":
          description: Suggestion not in an actionable state, or target row missing
  /problem-list-suggestions/{id}/reject:
    post:
      operationId: rejectProblemListSuggestion
      tags: [coding]
      summary: Reject a problem-list reconciliation action
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RejectProblemListSuggestionRequest" }
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProblemListSuggestion" }
        "404":
          description: Suggestion not found
        "409":
          description: Suggestion not in an actionable state
...
  /note-defaults:
    get:
      operationId: listNoteDefaults
      tags: [note-defaults]
      summary: List the signed-in provider's encounter defaults
      description: >-
        Per-provider "always apply" assumptions the AI bakes into every
        generated note (e.g. "14-point ROS negative unless stated").
      responses:
        "200":
          description: Note defaults
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/NoteDefault"
    post:
      operationId: createNoteDefault
      tags: [note-defaults]
      summary: Create a new note default
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNoteDefaultRequest"
      responses:
        "201":
          description: Note default created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NoteDefault"
  /note-defaults/{id}:
    patch:
      operationId: updateNoteDefault
      tags: [note-defaults]
      summary: Edit a note default
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateNoteDefaultRequest"
      responses:
        "200":
          description: Note default updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NoteDefault"
        "404":
          description: Note default not found (or not owned by caller)
    delete:
      operationId: deleteNoteDefault
      tags: [note-defaults]
      summary: Delete a note default
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Note default deleted
        "404":
          description: Note default not found (or not owned by caller)
...
components:
  schemas:
...
    NoteDefault:
      type: object
      required:
        - id
        - label
        - rule
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
        label:
          type: string
          minLength: 1
          maxLength: 120
        rule:
          type: string
          minLength: 1
          maxLength: 1000
          description: >-
            Imperative instruction the AI applies on every encounter
            unless the transcript contradicts it.
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateNoteDefaultRequest:
      type: object
      required: [label, rule]
      properties:
        label:
          type: string
          minLength: 1
          maxLength: 120
        rule:
          type: string
          minLength: 1
          maxLength: 1000
    UpdateNoteDefaultRequest:
      type: object
      description: >-
        Partial update. Any provided field replaces; omitted fields are
        untouched.
      properties:
        label:
          type: string
          minLength: 1
          maxLength: 120
        rule:
          type: string
          minLength: 1
          maxLength: 1000
    # ------------------------------------------------------------
    # Billing / Orders / Tasks — backfilled in Phase 28.
    # ------------------------------------------------------------
...
    SuggestionConfidence:
      type: string
      enum: [low, medium, high]
...
    BillingSupportingExcerpt:
      type: object
      required: [text]
      properties:
        text: { type: string }
        locationHint: { type: string }
...
    ProblemStatus:
      type: string
      enum: [active, stable, worsening, improving, resolved]
    ProblemEhrSource:
      type: string
      enum: [athena, epic, cerner, manual]
    ProblemSuggestionAction:
      type: string
      enum: [add, update_status, resolve, merge_duplicate, flag_uncertain]
    ProblemSuggestionStatus:
      type: string
      enum: [suggested, accepted, rejected, applied]
    PatientProblem:
      type: object
      required:
        - id
        - organizationId
        - patientId
        - code
        - description
        - status
        - ehrSource
        - createdAt
        - updatedAt
      properties:
        id: { type: string }
        organizationId: { type: string }
        patientId: { type: string }
        code: { type: string }
        description: { type: string }
        status: { $ref: "#/components/schemas/ProblemStatus" }
        onsetDate:
          type: ["string", "null"]
        ehrSource: { $ref: "#/components/schemas/ProblemEhrSource" }
        ehrResourceRef:
          type: ["string", "null"]
        syncedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ProblemListSuggestion:
      type: object
      required:
        - id
        - organizationId
        - patientId
        - encounterId
        - action
        - rationale
        - supportingExcerpts
        - confidence
        - status
        - appliedLocally
        - createdAt
        - updatedAt
      properties:
        id: { type: string }
        organizationId: { type: string }
        codingSessionId:
          type: ["string", "null"]
        patientId: { type: string }
        encounterId: { type: string }
        action: { $ref: "#/components/schemas/ProblemSuggestionAction" }
        targetProblemId:
          type: ["string", "null"]
        mergeFromProblemId:
          type: ["string", "null"]
        proposedCode:
          type: ["string", "null"]
        proposedDescription:
          type: ["string", "null"]
        proposedStatus:
          oneOf:
            - $ref: "#/components/schemas/ProblemStatus"
            - type: "null"
        rationale: { type: string }
        supportingExcerpts:
          type: array
          items: { $ref: "#/components/schemas/BillingSupportingExcerpt" }
        confidence: { $ref: "#/components/schemas/SuggestionConfidence" }
        status: { $ref: "#/components/schemas/ProblemSuggestionStatus" }
        statusNote:
          type: ["string", "null"]
        appliedLocally: { type: boolean }
        reviewedByUserId:
          type: ["string", "null"]
        reviewedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    AcceptProblemListSuggestionRequest:
      type: object
      properties:
        reason:
          type: string
          maxLength: 500
    RejectProblemListSuggestionRequest:
      type: object
      required: [reason]
      properties:
        reason:
          type: string
          minLength: 1
          maxLength: 500
...
    ReconcileResponse:
      type: object
      required: [data, problems, ehrHit]
      properties:
        data:
          type: array
          items: { $ref: "#/components/schemas/ProblemListSuggestion" }
        problems:
          type: array
          items: { $ref: "#/components/schemas/PatientProblem" }
        ehrHit:
          type: boolean
          description: True when Athena was queried (not mock / no ehrPatientId).
...